Updated the KotlinJS content project, the Android AAR library project and the Android example application project.
