+++ content/build.gradle
@@ -1,5 +1,5 @@
 buildscript {
-    ext.kotlin_version = '1.1.4-3'
+    ext.kotlin_version = '1.3.70'
     repositories {
         jcenter()
     }
@@ -9,7 +9,7 @@ buildscript {
 }
 
 group 'mockifer'
-version '2.0.0'
+version '3.0.0'
 
 apply plugin: 'kotlin2js'
 apply plugin: 'kotlin-dce-js'
@@ -29,8 +29,8 @@ compileKotlin2Js {
 
 // Need to keep the Mockifer methods that are called from the 'api.js' so they don't
 // get minified out of the build.
-runDceKotlinJs.keep "content_main.mockiferProcessRequest"
-runDceKotlinJs.keep "content_main.mockiferReset"
+runDceKotlinJs.keep "content.mockiferProcessRequest"
+runDceKotlinJs.keep "content.mockiferReset"
 
 repositories {
     jcenter()
@@ -112,15 +112,15 @@ build.doLast() {
     }
 
     copy {
-        from new File("${buildDir}/classes/kotlin/main/min/content_main.js")
+        from new File("${buildDir}/kotlin-js-min/main/content.js")
         into "${productDir}"
         rename { String fileName ->
-            fileName.replace('content_main.js', 'mockifer.js')
+            fileName.replace('content.js', 'mockifer.js')
         }
     }
 
     copy {
-        from new File("${buildDir}/classes/kotlin/main/min/kotlin.js")
+        from new File("${buildDir}/kotlin-js-min/main/kotlin.js")
         into "${productDir}"
     }
