+++ support/samples/sample-android/app/src/androidTest/java/io/github/marcelbraghetto/mockifer/example/android/MockiferTestRule.kt
@@ -1,0 +1,26 @@
+package io.github.marcelbraghetto.mockifer.example.android
+
+import android.net.Uri
+import androidx.test.platform.app.InstrumentationRegistry
+import androidx.appcompat.app.AppCompatActivity
+import androidx.test.rule.ActivityTestRule
+import io.github.marcelbraghetto.mockifer.Mockifer
+import io.github.marcelbraghetto.mockifer.example.android.main.BaseApplication
+
+class MockiferTestRule<T : AppCompatActivity>(activityClass: Class<T>) : ActivityTestRule<T>(activityClass, true, false) {
+    private companion object {
+        var mockiferInitialised = false
+    }
+
+    init {
+        if (!mockiferInitialised) {
+            mockiferInitialised = true
+
+            val uri = Uri.parse(BuildConfig.SERVER_BASE_URL)
+            val port = uri.port
+
+            Mockifer.installOnPort(BaseApplication.get(), InstrumentationRegistry.getInstrumentation().context, port)
+            Mockifer.setCommandUrl(uri.host, port)
+        }
+    }
+}
