--- support/samples/sample-android/app/src/androidTest/java/mockifer/example/RecyclerViewCountAssertion.kt
@@ -1,34 +1,0 @@
-package mockifer.example
-
-import android.support.test.espresso.NoMatchingViewException
-import android.support.test.espresso.ViewAssertion
-import android.support.v7.widget.RecyclerView
-import android.view.View
-import org.hamcrest.CoreMatchers.`is`
-import org.hamcrest.Matcher
-import org.hamcrest.MatcherAssert.assertThat
-
-internal class RecyclerViewCountAssertion private constructor(private val matcher: Matcher<Int>) : ViewAssertion {
-
-    override fun check(view: View, noViewFoundException: NoMatchingViewException?) {
-        if (noViewFoundException != null) {
-            throw noViewFoundException
-        }
-
-        val recyclerView = view as RecyclerView
-        val adapter = recyclerView.adapter
-        assertThat(adapter.itemCount, matcher)
-    }
-
-    companion object {
-
-        fun withItemCount(expectedCount: Int): RecyclerViewCountAssertion {
-            return withItemCount(`is`(expectedCount))
-        }
-
-        fun withItemCount(matcher: Matcher<Int>): RecyclerViewCountAssertion {
-            return RecyclerViewCountAssertion(matcher)
-        }
-    }
-}
-
