+++ support/samples/sample-android/app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest
-        package="mockifer.example.android"
+        package="io.github.marcelbraghetto.mockifer.example.android"
         xmlns:android="http://schemas.android.com/apk/res/android"
         >
