+++ support/samples/sample-android/app/src/main/java/io/github/marcelbraghetto/mockifer/example/android/cats/Cats.kt
@@ -1,0 +1,6 @@
+package io.github.marcelbraghetto.mockifer.example.android.cats
+
+import com.google.gson.annotations.SerializedName
+import io.github.marcelbraghetto.mockifer.example.android.cats.Cat
+
+class Cats(@SerializedName("cats") val cats: Array<Cat>)
