+++ support/samples/sample-android/app/src/main/java/io/github/marcelbraghetto/mockifer/example/android/main/BaseApplication.kt
@@ -1,0 +1,19 @@
+package io.github.marcelbraghetto.mockifer.example.android.main
+
+import android.annotation.SuppressLint
+import android.app.Application
+
+abstract class BaseApplication : Application() {
+    companion object {
+        @SuppressLint("StaticFieldLeak") private lateinit var instance: Application
+
+        fun get(): Application {
+            return instance
+        }
+    }
+
+    override fun onCreate() {
+        super.onCreate()
+        instance = this
+    }
+}
