--- support/samples/sample-android/app/src/main/java/mockifer/example/android/cats/Cats.kt
@@ -1,5 +1,0 @@
-package mockifer.example.android.cats
-
-import com.google.gson.annotations.SerializedName
-
-class Cats(@SerializedName("cats") val cats: Array<Cat>)
