--- support/samples/sample-android/app/src/main/java/mockifer/example/android/main/BaseApplication.kt
@@ -1,19 +1,0 @@
-package mockifer.example.android.main
-
-import android.annotation.SuppressLint
-import android.app.Application
-
-abstract class BaseApplication : Application() {
-    companion object {
-        @SuppressLint("StaticFieldLeak") private lateinit var instance: Application
-
-        fun get(): Application {
-            return instance
-        }
-    }
-
-    override fun onCreate() {
-        super.onCreate()
-        instance = this
-    }
-}
