+++ support/samples/sample-android/app/src/mock/java/io/github/marcelbraghetto/mockifer/example/android/MainApp.kt
@@ -1,0 +1,11 @@
+package io.github.marcelbraghetto.mockifer.example.android
+
+import io.github.marcelbraghetto.mockifer.Mockifer
+import io.github.marcelbraghetto.mockifer.example.android.main.BaseApplication
+
+class MainApp : BaseApplication() {
+    override fun onCreate() {
+        super.onCreate()
+        Mockifer.installOnPort(this, this.applicationContext, 8507)
+    }
+}
