+++ support/samples/sample-android/app/src/mock/java/io/github/marcelbraghetto/mockifer/example/android/MockiferUtil.kt
@@ -1,0 +1,13 @@
+package io.github.marcelbraghetto.mockifer.example.android
+
+import io.github.marcelbraghetto.mockifer.Mockifer
+
+object MockiferUtil {
+    fun pushMock(mockId: String) {
+        Mockifer.pushMock(mockId)
+    }
+
+    fun reset() {
+        Mockifer.reset()
+    }
+}
