--- support/samples/sample-android/app/src/mock/java/mockifer/example/android/MainApp.kt
@@ -1,11 +1,0 @@
-package mockifer.example.android
-
-import mockifer.Mockifer
-import mockifer.example.android.main.BaseApplication
-
-class MainApp : BaseApplication() {
-    override fun onCreate() {
-        super.onCreate()
-        Mockifer.installOnPort(this, this.applicationContext, 8507)
-    }
-}
