--- support/samples/sample-android/app/src/mock/java/mockifer/example/android/MockiferUtil.kt
@@ -1,13 +1,0 @@
-package mockifer.example.android
-
-import mockifer.Mockifer
-
-object MockiferUtil {
-    fun pushMock(mockId: String) {
-        Mockifer.pushMock(mockId)
-    }
-
-    fun reset() {
-        Mockifer.reset()
-    }
-}
