+++ support/samples/sample-android/app/src/normal/java/io/github/marcelbraghetto/mockifer/example/android/MainApp.kt
@@ -1,0 +1,5 @@
+package io.github.marcelbraghetto.mockifer.example.android
+
+import io.github.marcelbraghetto.mockifer.example.android.main.BaseApplication
+
+class MainApp : BaseApplication()
