+++ support/samples/sample-android/app/src/normal/java/io/github/marcelbraghetto/mockifer/example/android/MockiferUtil.kt
@@ -1,0 +1,12 @@
+package io.github.marcelbraghetto.mockifer.example.android
+
+object MockiferUtil {
+    @Suppress("UNUSED_PARAMETER")
+    fun pushMock(mockId: String) {
+        // Pushing mocks only supported in mock app
+    }
+
+    fun reset() {
+        // Reset only supported in mock app
+    }
+}
