--- support/samples/sample-android/app/src/normal/java/mockifer/example/android/MainApp.kt
@@ -1,5 +1,0 @@
-package mockifer.example.android
-
-import mockifer.example.android.main.BaseApplication
-
-class MainApp : BaseApplication()
