--- support/samples/sample-android/app/src/normal/java/mockifer/example/android/MockiferUtil.kt
@@ -1,11 +1,0 @@
-package mockifer.example.android
-
-object MockiferUtil {
-    fun pushMock(mockId: String) {
-        // Pushing mocks only supported in mock app
-    }
-
-    fun reset() {
-        // Reset only supported in mock app
-    }
-}
