+++ support/source/android/.gitignore
@@ -1,3 +1,4 @@
+.cxx
 *.iml
 .gradle
 /local.properties
